Add a pass for sorting optimizations from least to most general
This helps us ensure that we always match the least-general (aka most-specific)
optimization that we can for a particular instruction sequence.

For example, if we have both of these optimizations:

```lisp
(=> (imul $C $x)
    (imul_imm $C $x))

(=> (when (imul $C $x))
          (is-power-of-two $C))
    (ishl $x $C))
```

and we are matching `(imul 4 (..))`, then we want to apply the second
optimization, because it is more specific than the first.