Add the `peepmatic-automata` crate for finite-state transducer automata
A transducer is a type of automata that has not only an input that it accepts or
rejects, but also an output. While regular automata check whether an input
string is in the set that the automata accepts, a transducer maps the input
strings to values. A regular automata is sort of a compressed, immutable set,
and a transducer is sort of a compressed, immutable key-value dictionary. A
[trie] compresses a set of strings or map from a string to a value by sharing
prefixes of the input string. Automata and transducers can compress even better:
they can share both prefixes and suffixes. [*Index 1,600,000,000 Keys with
Automata and Rust* by Andrew Gallant (aka burntsushi)][burntsushi-blog-post] is
a top-notch introduction.

In `peepmatic` we will compile a set of linearized optimizations down into a
finite-state-transducer that simultaneously

* matches the lest-general and most-specific optimization that can apply to an
  instruction we're looking at, and

* builds up the right-hand side replacement instruction sequence for the match.

Because the transducer is building up the optimization's right-hand side
replacement, we are compressing and de-duplicating not just patterns we are
matching, but also their replacements.

We will do all of this incrementally: we cannot provide the full input sequence
for the automata all at once. We need to know which match operation to evaluate
next at each step as we walk through the automata. Therefore, each state in the
automata will have an associated match operation to evaluate, and its
transitions are switches on the resulting value.

This new `peepmatic-automata` crate is designed with these uses in mind.