Add an interpreter for applying peephole optimizations
This also renames `PeepholeOptimizer` into `PeepholeOptimizations` which is the
static, immutable transducer automaton compiled from the DSL. Then introduces
`PeepholeOptimizer` to be the mutable thing with dynamic allocations that
interprets the match operations against a generic `Instruction` and
`InstructionSet` and then applies the resulting actions.

In order to support actually applying optimizations, we needed to also assign
types and bit widths to optimizations and shepherd those assignments through
into compiled RHS actions. This way, we know how wide of a constant to create
for an `Action::MakeIntegerConst`, for example.